IntegrationTestExtension: Move to "apache"
As setting up the testing environment for the Apache client.
