+++ src/test/java/com/algolia/search/apache/index/ReplacingTest.java
@@ -1,6 +1,6 @@
 package com.algolia.search.apache.index;
 
-import com.algolia.search.integration.IntegrationTestExtension;
+import com.algolia.search.apache.IntegrationTestExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 @ExtendWith({IntegrationTestExtension.class})
