+++ src/test/java/com/algolia/search/apache/insights/InsightsTest.java
@@ -1,10 +1,10 @@
 package com.algolia.search.apache.insights;
 
-import static com.algolia.search.integration.IntegrationTestExtension.ALGOLIA_API_KEY_1;
-import static com.algolia.search.integration.IntegrationTestExtension.ALGOLIA_APPLICATION_ID_1;
+import static com.algolia.search.apache.IntegrationTestExtension.ALGOLIA_API_KEY_1;
+import static com.algolia.search.apache.IntegrationTestExtension.ALGOLIA_APPLICATION_ID_1;
 
 import com.algolia.search.InsightsClient;
-import com.algolia.search.integration.IntegrationTestExtension;
+import com.algolia.search.apache.IntegrationTestExtension;
 import java.io.IOException;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.TestInstance;
